algorithm tweaks

+++ analyzequotes.h
@@ -56,7 +56,7 @@ int maybeBuy(Quote yesterday, Quote today, TradeWeight *buyWeight, Portfolio *po
 int maybeSell(Quote yesterday, Quote today, TradeWeight *sellWeight, Portfolio *portfolio);
 TradeWeight * randomWeight(void);
 void randomizeWeight(TradeWeight *w);
-double score(Strategy s);
+int score(Strategy s);
 void bubbleSort(Strategy *s, int length);
 void generation(Strategy *s, int sCount, Quote *q, int qCount);
 void copyBytes(TradeWeight *twSource, TradeWeight *twDest);
